+++ src/app/core/core.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { AuthService } from './services/auth.service';
 
 @NgModule({
-  declarations: [],
   imports: [CommonModule],
 })
 export class CoreModule {