+++ src/app/core/models/credentialManagement.interface.ts
@@ -2,7 +2,7 @@
 import { CredentialMandatee } from "./credendentialMandatee.interface";
 import { Mandator } from "./madator.interface";
 import { Power } from "./power.interface";
-import {Signer} from "./credentialProcedure.interface";
+import { Signer } from "./credentialProcedure.interface";
 
 
 export interface CredentialManagement {