+++ src/app/core/services/auth.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { Router } from '@angular/router';
 import { map } from 'rxjs/operators';
 
 @Injectable({
@@ -10,18 +9,24 @@ import { map } from 'rxjs/operators';
 export class AuthService {
   public isAuthenticated$: Observable<boolean>;
   private isAuthenticatedSubject: BehaviorSubject<boolean>;
-  private userDataSubject: BehaviorSubject<any>;
+  private readonly userDataSubject: BehaviorSubject<any>;
   private tokenSubject: BehaviorSubject<string>;
   private mandatorSubject: BehaviorSubject<any>;
+  private signerSubject: BehaviorSubject<any>;
   private emailSubject: BehaviorSubject<string>;
-  private rol="";
-  public constructor(private oidcSecurityService: OidcSecurityService, private router: Router) {
+  private nameSubject: BehaviorSubject<string>;
+
+  private userPowers: any[] = [];
+
+  public constructor(private oidcSecurityService: OidcSecurityService) {
     this.isAuthenticatedSubject = new BehaviorSubject<boolean>(false);
     this.isAuthenticated$ = this.isAuthenticatedSubject.asObservable();
     this.userDataSubject = new BehaviorSubject<any>(null);
     this.tokenSubject = new BehaviorSubject<string>('');
     this.mandatorSubject = new BehaviorSubject<any>(null);
+    this.signerSubject = new BehaviorSubject<any>(null);
     this.emailSubject = new BehaviorSubject<string>('');
+    this.nameSubject = new BehaviorSubject<string>('');
 
     this.checkAuth().subscribe();
   }
@@ -29,38 +34,80 @@ export class AuthService {
   public checkAuth(): Observable<boolean> {
     return this.oidcSecurityService.checkAuth().pipe(map(({ isAuthenticated, userData, accessToken }) => {
       this.isAuthenticatedSubject.next(isAuthenticated);
+
       if (isAuthenticated) {
+        this.userPowers = this.extractUserPowers(userData);
         this.userDataSubject.next(userData);
-        this.tokenSubject.next(accessToken);
 
-        const base64Url = this.tokenSubject.getValue().split('.')[1];
-        const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-        const jsonPayload = decodeURIComponent(window.atob(base64).split('').map(function(c) {
-            return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-        }).join(''));
-    
-        this.rol= JSON.parse(jsonPayload)['resource_access']['account-console']['roles'][0]
-
-
+        const vcObject = this.extractVCFromUserData(userData)
 
         const mandator = {
-          organizationIdentifier: userData.organizationIdentifier,
-          organization: userData.organization,
-          commonName: userData.commonName,
-          emailAddress: userData.emailAddress,
-          serialNumber: userData.serialNumber,
-          country: userData.country
+          organizationIdentifier: vcObject.credentialSubject.mandate.mandator.organizationIdentifier,
+          organization: vcObject.credentialSubject.mandate.mandator.organization,
+          commonName: vcObject.credentialSubject.mandate.mandator.commonName,
+          emailAddress: vcObject.credentialSubject.mandate.mandator.emailAddress,
+          serialNumber: vcObject.credentialSubject.mandate.mandator.serialNumber,
+          country: vcObject.credentialSubject.mandate.mandator.country
         };
         this.mandatorSubject.next(mandator);
-        const emailName = userData.emailAddress.split('@')[0];
+
+        const  signer = {
+          organizationIdentifier: vcObject.credentialSubject.mandate.signer.organizationIdentifier,
+          organization: vcObject.credentialSubject.mandate.signer.organization,
+          commonName: vcObject.credentialSubject.mandate.signer.commonName,
+          emailAddress: vcObject.credentialSubject.mandate.signer.emailAddress,
+          serialNumber: vcObject.credentialSubject.mandate.signer.serialNumber,
+          country: vcObject.credentialSubject.mandate.signer.country
+        }
+        this.signerSubject.next(signer)
+
+        const emailName = vcObject.credentialSubject.mandate.mandator.emailAddress.split('@')[0];
+        const name = vcObject.credentialSubject.mandate.mandatee.first_name + ' ' + vcObject.credentialSubject.mandate.mandatee.last_name;
+
         this.emailSubject.next(emailName);
+        this.nameSubject.next(name);
       }
       return isAuthenticated;
     }));
   }
 
+  private extractVCFromUserData(userData: any) {
+    return userData.vc ? JSON.parse(userData.vc) : null;
+  }
+
+  private extractUserPowers(userData: any): any[] {
+    try {
+      const vcObject = this.extractVCFromUserData(userData)
+      return vcObject?.credentialSubject?.mandate?.power || [];
+    } catch (error) {
+      return [];
+    }
+  }
+
+
+  // POLICY: login_restriction_policy
+  public hasOnboardingExecutePower(): boolean {
+    return this.userPowers.some((power: any) => {
+      if (power.tmf_function === "Onboarding") {
+        const action = power.tmf_action;
+        return action === "Execute" || (Array.isArray(action) && action.includes("Execute"));
+      }
+      return false;
+    });
+  }
+
+  // POLICY: user_powers_restriction_policy
+  public hasIn2OrganizationIdentifier() : boolean {
+    const userData = this.userDataSubject.getValue();
+    return "VATES-B60645900" === userData.organizationIdentifier;
+  }
+
   public getMandator(): Observable<any> {
     return this.mandatorSubject.asObservable();
+  }
+
+  public getSigner(): Observable<any> {
+    return this.signerSubject.asObservable();
   }
 
   public login(): void {
@@ -70,11 +117,18 @@ export class AuthService {
   public handleLoginCallback(): void {
     this.oidcSecurityService.checkAuth().subscribe(({ isAuthenticated, userData, accessToken }) => {
       if (isAuthenticated) {
+
+        this.userPowers = this.extractUserPowers(userData);
+        const hasOnboardingPower = this.hasOnboardingExecutePower();
+
+        if (!hasOnboardingPower) {
+          this.logout();
+          return;
+        }
+
         this.isAuthenticatedSubject.next(true);
         this.userDataSubject.next(userData);
         this.tokenSubject.next(accessToken);
-      } else {
-        console.log('Authentication failed or not completed yet.');
       }
     });
   }
@@ -99,7 +153,9 @@ export class AuthService {
   public getToken(): Observable<string> {
     return this.tokenSubject.asObservable();
   }
-  public getRol(): any{
-    return this.rol;
+
+  public getName(): Observable<string> {
+    return this.nameSubject.asObservable()
   }
+
 }